Accept simple glyphs with zero contours
Some fonts in the wild were found where the contour count was zero but
the instruction length was present, leading it to be parsed as a
SimpleGlyph. Previously we returned an error when calculating the number
of coordinates because there was no last value in the
end_pts_of_contours array.