subset: Skip mappings that point at .notdef glyph
We don't need these mappings as they will map to .notdef through
omission. This change was made to handle the case where all format 4
sub-tables are required to include an entry that maps 0xFFFF to glyph 0
at the end of the segments. The presence of this was causing the code to
think that mapping was required and it forced a Unicode BMP compatible
sub-table format to be selected when a Mac Roman one would suffice.