Workaround fonts that contain invalid cmap format 4 segment
Apparently these fonts were created with a version of Fontographer with
a bug. It reaults in an id_range_offset entry with the value 0xFFFF,
which is almost certainly well beyond the end of the glyph id array.

Prior to this change Prince would fail to subset the
tests/w3c-svg-11se/resources/anglepoi.woff font used in the
tests/w3c-svg-11se/harness/htmlObject/masking-mask-01-b.html test.