Cache less in Travis
The default `cache: cargo` setting[1] caches $HOME/.cargo and
$TRAVIS_BUILD_DIR/target. This is producing a 38Gb cache archive, which
is slow to download and extract (~3 mins) and slow to compress and
upload. I think ultimately it's slowing builds down. A build is
currently taking about 20 mins. This change removes the `target`
directory from the cache to see how that affects things.

[1]: https://docs.travis-ci.com/user/caching/#rust-cargo-cache