subset: Allow Prince to omit the cmap table
There was a problem where PDF containing forms were not rendering
properly: all characters input into text fields came out as .notdef
boxes.

The change that triggered these behaviour was the inclusion of a cmap
table in subset fonts as part of Allsorts 0.9. Reading the PDF
specification revealed that when a TrueType font is embedded as a CID
font then the cmap should be omitted, as noted in the comment in
src/fonts/subsetting.m.