Handle cmap subsetting from Windows Symbol to MacRoman
When an input font has both a Windows Symbol cmap and MacRoman cmap we
favour Symbol over MacRoman. In our glyph mapping code, when the
encoding is Symbol we call legacy_symbol_char_code to try to map the
Unicode input character to a symbol char code.

Later on Prince does a check to see if the number of used glyphs in the
font will fit in a MacRoman table and whether the used glyphs are all
MacRoman compatible. If those checks pass Prince tells Allsorts to
generate MacRoman cmap sub-table.

When creating the subset cmap table the mappings to keep from the input
cmap are determined. Allsorts checks that the mapping is MacRoman
compatible. Notably a legacy_symbol_char_code style call was not made
meaning all glyphs were rejected as incompatible (since Symbol chars
typically occupy 0xF000..=0xF0FF, thus generating a MacRoman cmap that
maps everything to .notdef.

In this commit that is changed so that in the scenario of generating a
MacRoman cmap sub-table from Symbol sub-table the char codes are mapped
to Unicode.