fonts: Fix bug in read_packed_point_numbers
Point number runs are stored as deltas to the previous point number. For
the first run the previous number is 0 but for subsequent runs it's the
last point number of the preceding run. However, the code was resetting
to 0 at the start of each run, resulting in incorrect numbers for the
second and subsequent runs.