fonts: Fetch axis count of font upfront
A prior change in Prince uses an empty variation tuple to indicate
`None`/no variation tuple. This was chosen because `fvar.axis_count`
must be greater than zero for the font to be a functional variable font.
The OpenType specification says:

> A functional variable font must have an axisCount value that is
> greater than zero. If axisCount is zero, then the font is not functional
> as a variable font, and must be treated as a non-variable font; any
> variation-specific tables or data is ignored.

This change enforces this by making `Font::is_variable` only return true
if there is one or more axes, instead of just checking for the presence
of `fvar` and `gvar` like it did previously.

To lessen the impact of this change a ReadBinary impl for a new
FvarAxisCount struct is provided. It only reads and returns the axis
count from the `fvar` table instead of parsing the entire table.