Use default feature mask when shaping Thai and Lao text
The OT shaping documents specify that the LOCL and CCMP substitution
features should be applied in order to Thai and Lao text [1]. However,
some fonts in the wild (e.g. LN TILOK) make use of the LIGA feature
instead.

This change modifies the Thai/Lao shaper to use the default feature
mask, which includes the LOCL, CCMP, and LIGA features. It should be
noted that LOCL and CCMP are no longer applied in order, however this
(to my knowledge) was never an issue in older editions of Prince.

[1]: https://github.com/n8willis/opentype-shaping-documents/blob/master/opentype-shaping-thai-lao.md#stage-1-applying-the-language-substitution-features-from-gsub