fonts: Handle STAT axis value table format 3
Previously this was ignored as it was thought this this format did not
apply when generating sub-family names. That was wrong. It was only the
linkedValue field that could be ignored:

> Note: Applications are not required to use these style-linked mappings
> when implementing text formatting user interfaces. This data can be
> provided in a font for the benefit of applications that choose to do so.
> If a given application does not apply such style mappings for the given
> axis, then the linkedValue field is ignored.

When processing Inter with a wgth value of 700 and slnt value of 0 the
subfamily name was being generated as Bold Italic. With this change it
is Bold, as expected.