cff: Address error and warning from tx
When feeding a CFF font produced by subsetting a CFF2 font to [tx]
it emitted a warning:

tx: (cfr) two FSTypes (OS/2 value retained, CFF value removed)

and an error:

tx: (cfr) invalid OrigFontType value

The warning was addressed by only including the PostScript entry in the
Top DICT when the CFF font was to be standalone (not wrapped in
OpenType). The error was addressed by changing the OrigFontType from
/OpenType to /TrueType. Despite /OpenType appearing in the example in
the CFF specification it does not appear to be a valid value and is not
accepted by tx.

[tx]: http://adobe-type-tools.github.io/afdko/AFDKO-Overview.html#tx