fonts: Fix fake italic
In https://github.com/yeslogic/prince/pull/1419 the initial value of
font-style on an `@font-face` rule was changed from normal to auto. This
caused the FontDesc0 ^ font_desc_style = normal, condition to fail, thus
inhibiting the application of fake italic.

This commit addresses this by handling auto, as well as oblique with an
angle or range of angles. As part of this fix some changes were made to
how variable fonts are generated too. The macStyle and fsSelection
fields in the `head` and `OS/2` tables respectively are now updated to
reflect the weight, width, and style of the generated font instance.

This in turn helps Prince avoid apply fake italic to instances generated
from a font with a `slnt` axis.