shaping: make mark-to-mark positioning more selective
Reported in https://github.com/yeslogic/allsorts/issues/107

The selection of glyph pairs for mark-to-mark positioning was too eager.
For the text in question the ligature component position was not being
considered which was resulting in mark-to-mark positioning being applied
to marks 2 and 3 when it should not have. Considering the ligature
component position corrects this.

However, requiring the ligature component position to always match broke
some marks in Vietnamese. The solution was the introduction of a
ligature flag on RawGlyph so that a mark glyph can be tracked as a
ligature. This is used during GPOS to allow mark-to-mark positioning
between a pair of marks if one of them is a ligature despite having
differing ligature component positions.

This behaviour matches Harfbuzz:

https://github.com/harfbuzz/harfbuzz/blob/09a17a086c54e5c1b1aa674bba5b5449286e1480/src/OT/Layout/GPOS/MarkMarkPosFormat1.hh#L134-L138