fonts: Eliminate a potential source of API misuse in CDBT handling
If cbdt::lookup was called with a different glyph_id to that which was
supplied to find_strike it was possible that the function would panic
due to assuming that the same glyph_id would be passed.

The structure was changed to capture the glyph_id used when determining
the matching strike so that it did not have to be passed into the call
to retrieve the bitmap data.