fonts: Introduce ParseError::MissingTable
This allows errors that arise from missing tables to be clearer.
Additionally it permits code that previously reported a missing
`head` table to continue to do that now that `head` is eagerly
loaded in `Font`.

As part of this the error type for fontcode functions was re-worked to
return a struct.