morx: Fix unexpected glyph deletion during ligature substitution
The previous implementation makes the (reasonable) assumption that all
glyphs pushed onto the processing stack should be consumed to form a
ligature. However, this assumption is incorrect, and can result in too
many glyphs being deleted.

This fix rewrites a significant portion of the ligature substitution
logic, in part due to also wanting to avoid as much cloning as possible.