colr: Distinguish between error and no COLR data
It is valid for a glyph in a COLR font to not have any COLR data. In
this case we previously reported an error (although the follow on
behaviour of using the glyph as a regular outline was correct). Now
no error is reported in this case.

We also no longer report an error if a glyph does have COLR data,
but doesn't draw anything. This was the case for the space glyph in
the Nabla font.

Test expectations are updated as a result.