colr: Rework gradient sorting
Stop sorting in process_color_line as it's done in normalize_stops.

Rework normalize_stops to store the f32 representation of offsets so
that total_cmp can be used to sort the normalized stops. This makes
later binary searches on the stops that also use total_cmp legit.

Lastly check if the stops in radial gradients are sorted and if not
sort them.