gpos: Skip over input sequence when applying ChainContextPos
Previously when only incremented the current glyph by one when applying
a ChainContextPos rule. This commit alters the behaviour to match the
spec and progess over the input sequence, which might be more than one
glyph as was the case when positioning the 2 adult+2 child family glyph
in Segoe UI Emoji.