fonts: Match presentation if explicit variation selector is present
When mapping glyphs if an explicit text or emoji presentation selector
is supplied then set MatchingPresentation::Required. This handles a case
where font-family is sans-serif and a emoji with gender sign with
trailing VS16 is encountered. Without this change it comes out as the
base emoji followed by gender sign. With the change the sequence is
correctly mapped to the single gendered glyph.